+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 ### :house: Internal
 - `editor-common`
   - [1054](https://github.com/wix-incubator/rich-content/pull/1054) refactor calculateDiff
+- `general`
+  -[#1140](https://github.com/wix-incubator/rich-content/pull/1140) Dynamic import to 'react-tooltip'
 ### :bug: Bug Fix
 - `wrapper`
   - [#1132](https://github.com/wix-incubator/rich-content/pull/1132) palettes not working